% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovCoagBasePrediction.R
\docType{package}
\name{CovCoagBasePrediction}
\alias{CovCoagBasePrediction}
\title{CovCoagBasePrediction}
\description{
A package for running prediction model development studies using data in the OMOP CDM
}
